services/audio-processor/requirements.txt
fastapi
uvicorn
requests
google-cloud-speech
python-dotenv